Enforce JS MIME type checking for worker scripts
This is part of the general program of tightening MIME type checks on
scripts. d1c3679934f52450666750464effb97e14f555e0 introduced enforcement
for worker-imported scripts, and this commit completes the picture for
workers, leaving only "fetch a classic script" without the check.

Closes #3255.

This also makes some editorial tweaks to the script fetching algorithms
to align with modern conventions, while we're in the area.